+++ Master/CMeshField.cpp
@@ -143,11 +143,11 @@ void CMeshField::Buffer_Write(void)
 			{
 				if (w % 2 == 0)
 				{
-					Ipv[(m_VertexW + 1) * (2 * h) + w] = m_VertexW + (m_VertexW * h) + (w * 0.5);
+					Ipv[(m_VertexW + 1) * (2 * h) + w] = (WORD)(m_VertexW + (m_VertexW * h) + (w * 0.5));
 				}
 				else
 				{
-					Ipv[(m_VertexW + 1) * (2 * h) + w] = (m_VertexW * h) + ((w - 1) * 0.5);
+					Ipv[(m_VertexW + 1) * (2 * h) + w] = (WORD)((m_VertexW * h) + ((w - 1) * 0.5));
 				}
 			}
 		}
@@ -174,7 +174,7 @@ void CMeshField::Mesh_Field_Initialize(void)
 		for (int w = 0;w < m_VertexW;w++)
 		{
 			int n = w + m_VertexW * h;
-			m_Mfield[n].pos = D3DXVECTOR3(((float)m_meshWnum * 0.5f * -m_meshW) + (m_meshW * w), 0, ((float)m_meshHnum * 0.5 * m_meshW) + (-m_meshW * h));
+			m_Mfield[n].pos = D3DXVECTOR3(((float)m_meshWnum * 0.5f * -m_meshW) + (m_meshW * w), 0, ((float)m_meshHnum * 0.5f * m_meshW) + (-m_meshW * h));
 			m_Mfield[n].color = D3DCOLOR(0xffffffff);
 			m_Mfield[n].uv = D3DXVECTOR2(w * (1.0f / m_meshWnum), h * (1.0f / m_meshHnum));
 		}
